chore(pre-commit): update pre-commit hooks and organize configuration
Organizes the pre-commit configuration file by categorizing hooks
into sections for better readability and maintainability. Adds
comments to clarify the purpose of each section. Updates the
Ruff version to match the one specified in `pyproject.toml` for
consistency. Introduces a new section for project configuration
and dependency checks using Poetry, and a security check section
with Gitleaks.

docs: update developer and CLI documentation for consistency

Updates command names in `DEVELOPER.md` and CLI documentation
to reflect changes in the command naming convention. Ensures
consistency across documentation files by aligning command
descriptions and names with the actual implementation.

refactor(cli): rename CLI commands for clarity and consistency

Renames CLI commands to use hyphenated names for better clarity
and consistency. Moves documentation-related command
implementations to a separate module for better organization.
Removes redundant code and comments to streamline the CLI
implementation.
