+++ .pre-commit-config.yaml
@@ -1,9 +1,8 @@
-exclude: '(^\.archive|/typings)/'
-
 default_language_version:
   python: "3.13.2"
 
 repos:
+  # 1. Fast File Checks
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v5.0.0
     hooks:
@@ -17,26 +16,44 @@ repos:
       - id: validate-pyproject
         additional_dependencies: ["validate-pyproject-schema-store[all]"]
 
+    # 2. Code Upgraders/Modifiers
+  - repo: https://github.com/asottile/pyupgrade
+    rev: v3.19.1
+    hooks:
+      - id: pyupgrade
+        args: ["--py313-plus"]
+
   - repo: https://github.com/asottile/add-trailing-comma
     rev: v3.1.0
     hooks:
       - id: add-trailing-comma
 
-  - repo: https://github.com/asottile/pyupgrade
-    rev: v3.19.1
-    hooks:
-      - id: pyupgrade
-
+    # 3. Main Linter (with auto-fix)
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.11.2
+    # Ruff version should match the one in pyproject.toml
+    rev: v0.11.4 # Use the same Ruff version tag as formatter
     hooks:
-      # Run the linter.
       - id: ruff
         args: [--fix]
-      # Run the formatter.
+
+    # 4. Main Formatter (after linting/fixing)
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    # Ruff version should match the one in pyproject.toml
+    rev: v0.11.4
+    hooks:
       - id: ruff-format
 
+    # 5. Project Config / Dependency Checks
+  - repo: https://github.com/python-poetry/poetry
+    rev: 2.1.2 # Use the latest tag from the repo
+    hooks:
+      - id: poetry-check
+      - id: poetry-lock
+
+    # 6. Security Check
   - repo: https://github.com/gitleaks/gitleaks
-    rev: v8.24.2
+    rev: v8.24.2 # Use the latest tag from the repo
     hooks:
       - id: gitleaks
+
+exclude: '(^\.archive|/typings)/'
