+++ pyproject.toml
@@ -8,16 +8,14 @@ readme = "README.md"
 urls = { repository = "https://github.com/allthingslinux/tux" }
 dynamic = ["dependencies"]
 
+[project.scripts]
+tux = "tux.cli:main"
+
 [build-system]
 requires = ["poetry-core>=2.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-name = "tux"
-version = "0.1.0"
-description = "Tux is an all in one bot for the All Things Linux discord server."
-authors = ["All Things Linux <tux@allthingslinux.org>"]
-readme = "README.md"
 packages = [{ include = "tux" }]
 
 [tool.poetry.dependencies]
@@ -53,7 +51,6 @@ arrow = "^1.3.0"
 click = "^8.1.8"
 single-source = "^0.4.0"
 
-
 [tool.poetry.group.dev.dependencies]
 pre-commit = ">=4.0.0"
 pyright = ">=1.1.358"
@@ -78,9 +75,6 @@ griffe-generics = "^1.0.13"
 griffe-inherited-method-crossrefs = "^0.0.1.4"
 griffe-inherited-docstrings = "^1.1.1"
 mkdocs-api-autonav = "^0.2.0"
-
-[tool.poetry.scripts]
-tux = "tux.cli:main"
 
 [tool.ruff]
 cache-dir = ".cache/ruff"
