+++ tux/cli/dev.py
@@ -1,15 +1,14 @@
-"""Development commands for the Tux CLI."""
+"""Development tools and utilities for Tux."""
 
 from tux.cli.core import command_registration_decorator, create_group
 
-# Create the development command group
-dev_group = create_group("dev", "Development tools and commands")
+# Create the dev command group
+dev_group = create_group("dev", "Development tools")
 
 
 @command_registration_decorator(dev_group, name="lint")
 def lint() -> int:
-    """Run linters on the codebase."""
-
+    """Run linting with Ruff."""
     from tux.cli.impl.dev import lint as run_lint
 
     return run_lint()
@@ -17,8 +16,7 @@ def lint() -> int:
 
 @command_registration_decorator(dev_group, name="lint-fix")
 def lint_fix() -> int:
-    """Run linters and apply fixes."""
-
+    """Run linting with Ruff and apply fixes."""
     from tux.cli.impl.dev import lint_fix as run_lint_fix
 
     return run_lint_fix()
@@ -26,26 +24,23 @@ def lint_fix() -> int:
 
 @command_registration_decorator(dev_group, name="format")
 def format_code() -> int:
-    """Run code formatters."""
+    """Format code with Ruff."""
+    from tux.cli.impl.dev import format_code as run_format_code
 
-    from tux.cli.impl.dev import format_code as run_format
-
-    return run_format()
-
-
-@command_registration_decorator(dev_group, name="typecheck")
-def typecheck() -> int:
-    """Check types with pyright."""
-
-    from tux.cli.impl.dev import typecheck as run_typecheck
-
-    return run_typecheck()
+    return run_format_code()
 
 
-@command_registration_decorator(dev_group, name="check")
+@command_registration_decorator(dev_group, name="type-check")
+def type_check() -> int:
+    """Check types with Pyright."""
+    from tux.cli.impl.dev import type_check as run_type_check
+
+    return run_type_check()
+
+
+@command_registration_decorator(dev_group, name="pre-commit")
 def check() -> int:
-    """Run all checks (lint, format, typecheck)."""
-
-    from tux.cli.impl.dev import check as run_check
+    """Run pre-commit checks."""
+    from tux.cli.impl.dev import pre_commit as run_check
 
     return run_check()
