+++ tux/cli/impl/__init__.py
@@ -5,11 +5,11 @@ This package contains the actual implementations of CLI commands.
 
 from tux.cli.impl.core import run_command
 from tux.cli.impl.database import db_generate, db_migrate, db_pull, db_push, db_reset
-from tux.cli.impl.dev import check, docs, docs_build, format_code, lint, lint_fix, typecheck
+from tux.cli.impl.dev import format_code, lint, lint_fix, pre_commit, type_check
 from tux.cli.impl.docker import docker_build, docker_down, docker_exec, docker_logs, docker_ps, docker_up
+from tux.cli.impl.docs import docs_build, docs_serve
 
 __all__ = [
-    "check",
     "db_generate",
     "db_migrate",
     "db_pull",
@@ -21,11 +21,12 @@ __all__ = [
     "docker_logs",
     "docker_ps",
     "docker_up",
-    "docs",
     "docs_build",
+    "docs_serve",
     "format_code",
     "lint",
     "lint_fix",
+    "pre_commit",
     "run_command",
-    "typecheck",
+    "type_check",
 ]
